Add appointment email notifications
