+++ src/app/api/doctor/appointments/[id]/route.ts
@@ -15,168 +15,15 @@ import {
 } from "@/lib/time";
 import { sendEmail } from "@/lib/email";
 import { EMAIL_VERIFICATION_TOKEN_TYPE } from "@/lib/email-verification";
+import {
+    buildMoveEmail,
+    buildStatusEmail,
+    formatDoctorName,
+    formatPatientName,
+    getPatientEmail,
+} from "@/lib/appointment-email";
 
-const EMAIL_BACKGROUND_COLOR = "#f0fdf4";
-const EMAIL_BORDER_COLOR = "#bbf7d0";
-const EMAIL_TEXT_COLOR = "#065f46";
-const EMAIL_ACCENT_COLOR = "#047857";
 
-function escapeHtml(input: string) {
-    return input
-        .replace(/&/g, "&amp;")
-        .replace(/</g, "&lt;")
-        .replace(/>/g, "&gt;")
-        .replace(/"/g, "&quot;")
-        .replace(/'/g, "&#39;");
-}
-
-function formatPatientName(patient: {
-    username: string;
-    student: { fname: string | null; lname: string | null } | null;
-    employee: { fname: string | null; lname: string | null } | null;
-}) {
-    if (patient.student?.fname && patient.student?.lname)
-        return `${patient.student.fname} ${patient.student.lname}`;
-    if (patient.employee?.fname && patient.employee?.lname)
-        return `${patient.employee.fname} ${patient.employee.lname}`;
-    return patient.username;
-}
-
-function getPatientEmail(patient: {
-    username: string;
-    student: { email: string | null } | null;
-    employee: { email: string | null } | null;
-    passwordResetTokens: { contact: string }[];
-}) {
-    const verifiedContacts = new Set(
-        patient.passwordResetTokens.map((token) => token.contact.toLowerCase()),
-    );
-
-    const studentEmail = patient.student?.email?.trim() ?? "";
-    if (studentEmail && verifiedContacts.has(studentEmail.toLowerCase())) {
-        return studentEmail;
-    }
-
-    const employeeEmail = patient.employee?.email?.trim() ?? "";
-    if (employeeEmail && verifiedContacts.has(employeeEmail.toLowerCase())) {
-        return employeeEmail;
-    }
-
-    return patient.username.includes("@") ? patient.username : "";
-}
-
-function formatDoctorName(doctor: {
-    username: string;
-    employee: { fname: string | null; lname: string | null } | null;
-}) {
-    if (doctor.employee?.fname && doctor.employee?.lname)
-        return `Dr. ${doctor.employee.fname} ${doctor.employee.lname}`;
-    return doctor.username.startsWith("Dr.")
-        ? doctor.username
-        : `Dr. ${doctor.username}`;
-}
-
-function buildStatusEmail({
-    status,
-    patientName,
-    clinicName,
-    schedule,
-    doctorName,
-    cancelReason,
-}: {
-    status: AppointmentStatus;
-    patientName: string;
-    clinicName: string;
-    schedule: string;
-    doctorName: string;
-    cancelReason?: string | null;
-}): { subject: string; html: string; text: string } | null {
-    const safeName = escapeHtml(patientName);
-    const safeClinic = escapeHtml(clinicName);
-    const safeSchedule = escapeHtml(schedule);
-    const safeDoctor = escapeHtml(doctorName);
-    const safeReason = cancelReason ? escapeHtml(cancelReason) : null;
-
-    const rows = [
-        { label: "Clinic", value: safeClinic },
-        { label: "Doctor", value: safeDoctor },
-        { label: "Schedule", value: safeSchedule },
-    ];
-
-    const textRows: Array<[string, string]> = [
-        ["Clinic", clinicName],
-        ["Doctor", doctorName],
-        ["Schedule", schedule],
-    ];
-
-    let heading = "";
-    let intro = "";
-    let outro = `Thank you,<br/>${safeDoctor}<br/>HNU Clinic`;
-    let subject = "";
-    let textIntro = "";
-    let textOutro = `Thank you,\n${doctorName}\nHNU Clinic`;
-
-    switch (status) {
-        case AppointmentStatus.Approved:
-            heading = "Appointment Approved";
-            intro = `<span style="color: ${EMAIL_ACCENT_COLOR}; font-weight: 600;">Good news, <strong style="color: inherit;">${safeName}</strong>! <strong style="color: inherit;">${safeDoctor}</strong> has approved your appointment.</span>`;
-            subject = "Your appointment has been approved";
-            textIntro = `Good news, ${patientName}! ${doctorName} has approved your appointment.`;
-            break;
-        case AppointmentStatus.Cancelled:
-            heading = "Appointment Cancelled";
-            intro = `<span style="color: ${EMAIL_ACCENT_COLOR}; font-weight: 600;">Hello <strong style="color: inherit;">${safeName}</strong>, <strong style="color: inherit;">${safeDoctor}</strong> has cancelled your appointment.</span>`;
-            subject = "Your appointment has been cancelled";
-            if (safeReason) rows.push({ label: "Cancellation Reason", value: safeReason });
-            if (cancelReason) textRows.push(["Cancellation Reason", cancelReason]);
-            outro = `If you still need assistance, please book another schedule through the patient portal.<br/><br/>Thank you,<br/>${safeDoctor}<br/>HNU Clinic`;
-            textIntro = `Hello ${patientName}, ${doctorName} has cancelled your appointment.`;
-            textOutro = `If you still need assistance, please book another schedule through the patient portal.\n\nThank you,\n${doctorName}\nHNU Clinic`;
-            break;
-        case AppointmentStatus.Completed:
-            heading = "Appointment Completed";
-            intro = `<span style="color: ${EMAIL_ACCENT_COLOR}; font-weight: 600;">Hello <strong style="color: inherit;">${safeName}</strong>, <strong style="color: inherit;">${safeDoctor}</strong> has marked your appointment as completed.</span>`;
-            subject = "Your appointment has been completed";
-            outro = `We hope you had a good visit. You can review your consultation notes and next steps inside the patient portal.<br/><br/>Thank you,<br/>${safeDoctor}<br/>HNU Clinic`;
-            textIntro = `Hello ${patientName}, ${doctorName} has marked your appointment as completed.`;
-            textOutro = `We hope you had a good visit. You can review your consultation notes and next steps inside the patient portal.\n\nThank you,\n${doctorName}\nHNU Clinic`;
-            break;
-        default:
-            return null;
-    }
-
-    const rowHtml = rows
-        .map(
-            (row) => `
-        <tr>
-          <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR}; font-weight: 600;">${row.label}</td>
-          <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR};">${row.value}</td>
-        </tr>`
-        )
-        .join("");
-
-    const html = `
-    <div style="font-family: 'Segoe UI', Roboto, Helvetica, Arial, sans-serif; background-color: ${EMAIL_BACKGROUND_COLOR}; padding: 24px; border-radius: 16px; border: 1px solid ${EMAIL_BORDER_COLOR}; color: ${EMAIL_TEXT_COLOR};">
-      <h2 style="margin-top: 0; color: ${EMAIL_ACCENT_COLOR};">${heading}</h2>
-      <p>${intro}</p>
-      <table style="width: 100%; border-collapse: collapse; margin: 16px 0;">
-        <tbody>${rowHtml}</tbody>
-      </table>
-      <p style="margin-bottom: 0;">${outro}</p>
-    </div>`;
-
-    const text = [
-        textIntro,
-        "",
-        ...textRows.map(([label, value]) => `${label}: ${value}`),
-        "",
-        textOutro,
-    ]
-        .join("\n")
-        .trim();
-
-    return { subject, html, text };
-}
 function shapeResponse(appointment: {
     appointment_id: string;
     appointment_timestart: Date;
@@ -412,60 +259,21 @@ export async function PATCH(
             const doctorName = formatDoctorName(updated.doctor);
 
             if (patientEmail) {
-                const oldSchedule = formatManilaDateTime(appointment.appointment_timestart);
-                const newSchedule = formatManilaDateTime(updated.appointment_timestart);
-                const html = `
-                    <div style="font-family: 'Segoe UI', Roboto, Helvetica, Arial, sans-serif; background-color: ${EMAIL_BACKGROUND_COLOR}; padding: 24px; border-radius: 16px; border: 1px solid ${EMAIL_BORDER_COLOR}; color: ${EMAIL_TEXT_COLOR};">
-                        <h2 style="margin-top: 0; color: ${EMAIL_ACCENT_COLOR};">Appointment Update</h2>
-                        <p style="color: ${EMAIL_ACCENT_COLOR}; font-weight: 600;">Hello <strong style="color: inherit;">${escapeHtml(patientName)}</strong>, <strong style="color: inherit;">${escapeHtml(doctorName)}</strong> has updated your appointment.</p>
-                        <p>Your visit at <strong>${escapeHtml(updated.clinic.clinic_name)}</strong> has been moved.</p>
-                        <table style="width: 100%; border-collapse: collapse; margin: 16px 0;">
-                            <tbody>
-                                <tr>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR}; font-weight: 600;">Doctor</td>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR};">${escapeHtml(doctorName)}</td>
-                                </tr>
-                                <tr>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR}; font-weight: 600;">Previous Schedule</td>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR};">${escapeHtml(oldSchedule)}</td>
-                                </tr>
-                                <tr>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR}; font-weight: 600;">New Schedule</td>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR};">${escapeHtml(newSchedule)}</td>
-                                </tr>
-                                <tr>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR}; font-weight: 600;">Doctor's Note</td>
-                                    <td style="padding: 8px; border: 1px solid ${EMAIL_BORDER_COLOR};">${escapeHtml(trimmedReason)}</td>
-                                </tr>
-                            </tbody>
-                        </table>
-                        <p>Please log in to your patient portal if you need to reschedule again or have any questions.</p>
-                        <p style="margin-bottom: 0;">Thank you,<br/>${escapeHtml(doctorName)}<br/>HNU Clinic</p>
-                    </div>
-                `;
-
-                const text = [
-                    `Hello ${patientName},`,
-                    "",
-                    `${doctorName} has updated your appointment at ${updated.clinic.clinic_name}.`,
-                    "",
-                    `Previous Schedule: ${oldSchedule ?? "Unavailable"}`,
-                    `New Schedule: ${newSchedule ?? "Unavailable"}`,
-                    `Doctor's Note: ${trimmedReason}`,
-                    "",
-                    "Please log in to your patient portal if you need to reschedule again or have any questions.",
-                    "",
-                    "Thank you,",
+                const moveEmail = buildMoveEmail({
+                    patientName,
                     doctorName,
-                    "HNU Clinic",
-                ].join("\n");
+                    clinicName: updated.clinic.clinic_name,
+                    oldSchedule: formatManilaDateTime(appointment.appointment_timestart),
+                    newSchedule: formatManilaDateTime(updated.appointment_timestart),
+                    reason: trimmedReason,
+                });
 
                 try {
                     await sendEmail({
                         to: patientEmail,
-                        subject: "Your appointment has been moved",
-                        html,
-                        text,
+                        subject: moveEmail.subject,
+                        html: moveEmail.html,
+                        text: moveEmail.text,
                     });
                 } catch (emailErr) {
                     console.error("[PATCH /api/doctor/appointments/:id] email error", emailErr);
